Refact nextauth handlers
